test: 리팩토링에 대응해 테스트 코드 수정
